requirements.txt
# Core Data Science Libraries
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.11.1

# Machine Learning & Clustering
scikit-learn>=1.3.0
xgboost>=1.7.6

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Statistical Analysis
statsmodels>=0.14.0

# Excel/File Processing
openpyxl>=3.1.2
xlrd>=2.0.1

# Web Framework for Dashboard
dash>=2.11.1
dash-bootstrap-components>=1.4.1

# Report Generation
jinja2>=3.1.2
weasyprint>=59.0

# Configuration
pyyaml>=6.0.1

# Jupyter for Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Progress Bars
tqdm>=4.65.0

# Date/Time Processing
python-dateutil>=2.8.2